General restructure; Notebook revival; data_loader.py has been set up; Grad cams have been revised and updated
The full notebook is now back up and running. The saved models were using methods that were so old, they cannot be loaded. I will be doing some model training it seems.
The grad cams were not working initially and required a little change. After making the change it turned out that the functions work but are depreciated. I got an updated version working using a matplotlib built in function. I still have a lot of cleanup to do.